lab8/requirements.txt
lxml==4.6.3
requests==2.26.0
beautifulsoup4==4.10.0
Pillow==8.3.2
PyPDF2==1.26.0
